Fix the install and build commands.
This commit fixes 2 issues:

1. The `install`, `build`, and `build-nix` commands always failed,
because they tried to invoke the Nix-generated install and build
scripts by executing the paths of their grandparent directories, not
the paths of the scripts themselves.

2. These failures were not effectively communicated to the user,
because the `runCommand` action ignored errors, and because the bash
scripts that invoke these commands were not run with the `-e` flag. If
you weren't paying careful attention to the warning displayed on
stdout, or if you were running these commands as part of a CI
script (e.g., in the test/tests.bash script), you likely wouldn't
notice that the commands had failed.