Fix generated Nix scripts to properly fail on error.
One more fix related to the previous commits: the Nix scripts are
generated via `writeShellScriptBin` and therefore already start with
the bash shebang, so the bash shebang that we added in the body of the
script was a) redundant and b) ignored, thereby making the `-e` flag
to bash ineffective.

Now we simply `set -e` on the first line of the script.